Move NoPrint from environment into configuration (#90)
