+++ src/math/linear_solver_implementation.hpp
@@ -9,7 +9,7 @@
 /**
  * @file   linear_solver_implementation.hpp
  * @author William A. Perkins
- * @date   2015-08-18 13:40:52 d3g096
+ * @date   2022-10-05 08:52:17 d3g096
  * 
  * @brief  
  * 
@@ -29,7 +29,7 @@
 #include <gridpack/math/linear_solver_interface.hpp>
 #include <gridpack/parallel/distributed.hpp>
 #include <gridpack/utilities/uncopyable.hpp>
-#include <gridpack/environment/no_print.hpp>
+#include <gridpack/configuration/no_print.hpp>
 #include <gridpack/configuration/configurable.hpp>
 
 namespace gridpack {
